+++ packages/terra-framework-docs/CHANGELOG.md
@@ -13,6 +13,8 @@
   * Migrated docs from terra-navigation-prompt.
   * Migrate docs from terra-navigation-layout
   * Migrate docs from terra-menu
+  * Migrate docs from terra-notification-dialog.
+
 
 ## 1.7.0 - (March 8, 2022)